+++ docs/source/changelog.rst
@@ -55,7 +55,7 @@ Release 1.0
 
 * ``Cordformation`` and node identity generation
   * Cordform may not specify a ``NetworkMapNode``, when that happens during ``DeployNodes`` the following happens:
-    1. Each node is started and its signed serialized NodeInfo is written to disk in the node folder.
+    1. Each node is started and its signed serialized NodeInfo is written to disk in the node base directory.
     2. Every serialized ``NodeInfo`` above is copied in every other node "additional-node-info" folder under the NodeInfo folder.
   * Nodes read all the nodes stored in ``additional-node-info`` when the ``NetworkMapService`` starts up.